requirements.txt
PyQt6==6.8.1
google-generativeai>=0.6.0
sounddevice>=0.4.5
cryptography>=41.0.0
python-dotenv>=1.0.0
requests>=2.30.0
numpy>=1.24.0
